wip: reorganize ndc-test into modules
